Background Image - green sky,migrate create to root , removed create root
