+++ components/Layout/BaseLayout/BaseLayout.tsx
@@ -3,7 +3,7 @@ import { ILayout } from "../types"
 
 const BaseLayout = ({ children }: ILayout) => (
   <div
-    className="w-screen h-screen bg-[url('/images/Layout/background.png')] bg-contain
+    className="w-screen h-screen bg-[url('/images/Layout/background-green.png')] bg-contain
     flex justify-center overflow-hidden"
   >
     {children}
